TBAPP/back/requirements.txt
asn1crypto==1.0.1
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
python-dotenv==0.15.0
gunicorn==20.0.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycparser==2.19
Werkzeug==0.16.0
Flask_Cors==3.0.8
Flask_SQLAlchemy==2.4.1
psycopg2==2.8.4
PyJWT==1.7.1
six==1.13.0
SQLAlchemy==1.3.11


